Add systemd --user support
Use systemd --user to spawn the session.
It's enabled by default, but it can be turned off.

When enabled, the session modules are not used and in their stead
a systemd target is started. Autostart applications are started by a
generator that creates the units on the fly.

liri-daemon is totally useless now, and it's turned into a program that
loads the module and registers a D-Bus service. This is then used as a
systemd service, for the liri-daemon target.

Closes: #4