Integrate with logind
Lock/unlock session using logind, lock before sleep and shutdown when
the system is going off.

This code was originally written for shell, but it belongs here.

Keep the modular architecture so that in the future we can easily
add new backends (for example ConsoleKit, or any other solution that
is alternative to systemd).

Closes: #11