Run applications with systemd-run
When using systemd, run commands and programs with systemd otherwise
keep spawning processes the way we used to do.

Either use systemd-run for D-Bus activation for D-Bus activatable
apps (which uses systemd under the hood) or systemd-run.

We separated the process launcher from the session manager to make sure
the environment variables didn't get mixed up, but that problem can be
avoided with some care.

Merge the process launcher with the session manager and make sure the
environment to launch the shell don't get mixed with the environment
used to launch apps.

This has two benefits: the D-Bus API is still the same, but we have
one less daemon to take care of.

Since the autostart module uses the process launcher, enable it when
systemd support is in use. The other session modules are disabled.

Ignore autostart entries with X-GNOME-HiddenUnderSystemd when using systemd.

Some programs ship an autostart entry and a unit, an example is pulseaudio.
In order to prevent running such services twice, we put a hidden entry
in a special directory $datadir/liri-session/systemd-user/autostart.

At this point we don't need the autostart generator anymore.

Closes: #6